Some updates
Added glyphicons to indicate what they do. Added target="_blank" to links that needed it. Gave escape.html the project id to fit in with the other project pages. Changed "my email" to my actual email on the home page. Added the download and video links to mmshow2016.html (finally!).